Refactored tests + SDRAM class
Moved tests to tests.cpp. Most of these will no longer
be useful with the change from a memory-mapped
audio core to a streaming audio core. I will keep them
For now.

SDRAM wrapper class has not been tested yet.
