+++ firmware_391/sdram.h
@@ -1,0 +1,21 @@
+/*
+ * sdram.h
+ *
+ *  Created on: Mar 22, 2021
+ *      Author: Tyler
+ */
+
+#ifndef SDRAM_H_
+#define SDRAM_H_
+
+class SDRAM {
+private:
+	volatile short *address;
+public:
+	SDRAM(unsigned int address);
+	short get(unsigned int offset);
+	void put(unsigned int offset, short value);
+};
+
+
+#endif /* SDRAM_H_ */
